ai-agent-service/requirements.txt
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0

# LLM (OpenAI-compatible API)
openai>=1.30.0

# Optional: Postgres + vectors (install when you wire memory)
# psycopg2-binary>=2.9.9
# sqlalchemy>=2.0.0
# faiss-cpu>=1.8.0
